requirements.txt
# CGMSV Launcher Dependencies
# Core dependencies for CGMSV Launcher v1.0.0

# Process management
psutil>=5.8.0

# Configuration file parsing
PyYAML>=6.0

# Executable building (optional, for development)
pyinstaller>=5.0.0 